Promote packages
Adds plugin code to promote packages from one repository to another.

To test, we actually push a test container to a dedicated testing
repository, promote it to another testing repository, and then verify
that the package was indeed moved. (These repositories are configured
to automatically delete packages older than 1 day).

Extensive documentation is available in the README, as well as in the
various unit tests.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>