Add merge pipeline
To ensure that we only build a Cloudsmith CLI container when
necessary, we pull out any testing container declarations in
`docker-bake.hcl` (this lets our diffing logic be totally matched up
with reality; if anything changes in `docker-bake.hcl` or
`Dockerfile`, we've got a new container.

The merge pipeline itself has the logic for building the container, if
necessary, and then testing the plugin using that new container. If
all goes well, we'll promote the container to our `releases`
repository (using the current official release of the plugin).

Signed-off-by: Christopher Maier <chris@graplsecurity.com>