Add 'action' to plugin
This allows us to promote by either copy or move (previously, it was
only move). This offers more flexibility in how we handle our
promotions.

Of particular note is the fact that the upload container -> promote
container -> verify promotion logic used in our pipelines to verify
the action of the pipeline has been extracted into a dynamic pipeline,
which allows us to forego a bunch of similar-but-not-quite-the-same
copy-and-paste of a lot of job descriptions across multiple pipelines.

The behavior of the new `action` parameter is documented in the
README.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>